lib: fix push on prematurely terminated connection
When a SPDY connection is terminated prematurely, it is possible for
user-code to still hold references to the request and response objects
and perform operations on them, including creating a new push stream.
This condition can trigger an uncaught exception when the underlying
connection has been closed.

This commit fixes an issue in the Response class by eliminating the need
to read the name of the connected host using the parser attached to the
socket. Instead, it uses a reference to its corresponding request object
(now setup in server.js when a "request" event is fired) to read the
HTTP Host header.