handle: fix misfiring callback
Streams can be ended by calling the `.close()` or `.shutdown()` methods
on the Handle instances. Internally, both call the `.end()` method to
end the stream. However, only the `.close()` method takes a callback
that eventually frees the HttpParser instance allocated to the stream
socket.

In the event that the `.shutdown()` method is called before calling the
`.close()` method, the second call to `.end()` never fires the callback,
resulting in the leaking of HttpParser instances along with all its
references.

This commit fixes the issue by directly executing the callback provided
to the `.close()` method in the event the `.shutdown()` method was
called before the `.close()` method of the Handle instance.